+++ Mile.Helpers/Mile.Helpers.cpp
@@ -10,6 +10,33 @@
 
 #include "Mile.Helpers.h"
 
+namespace
+{
+    static bool IsPrivatePerMonitorSupportExtensionApplicable()
+    {
+        static bool CachedResult = ([]() -> bool
+        {
+            // The private Per-Monitor DPI Awareness support extension is
+            // Windows 10 only.
+            if (!::MileIsWindowsVersionAtLeast(10, 0, 0))
+            {
+                return false;
+            }
+
+            // We don't need the private Per-Monitor DPI Awareness support
+            // extension if the Per-Monitor (V2) DPI Awareness exists.
+            if (::MileIsWindowsVersionAtLeast(10, 0, 14986))
+            {
+                return false;
+            }
+
+            return true;
+        }());
+
+        return CachedResult;
+    }
+}
+
 EXTERN_C BOOL WINAPI MileIsWindowsVersionAtLeast(
     _In_ DWORD Major,
     _In_ DWORD Minor,
@@ -33,4 +60,32 @@ EXTERN_C BOOL WINAPI MileIsWindowsVersionAtLeast(
                 VER_GREATER_EQUAL),
             VER_BUILDNUMBER,
             VER_GREATER_EQUAL));
+}
+
+EXTERN_C BOOL WINAPI MileEnablePerMonitorDialogScaling()
+{
+    if (!::IsPrivatePerMonitorSupportExtensionApplicable())
+    {
+        return FALSE;
+    }
+
+    HMODULE ModuleHandle = ::GetModuleHandleW(L"user32.dll");
+    if (!ModuleHandle)
+    {
+        return FALSE;
+    }
+
+    // Reference: https://github.com/microsoft/terminal/blob
+    //            /9b92986b49bed8cc41fde4d6ef080921c41e6d9e
+    //            /src/interactivity/win32/windowdpiapi.cpp#L24
+    typedef BOOL(WINAPI* ProcType)();
+
+    ProcType ProcAddress = reinterpret_cast<ProcType>(
+        ::GetProcAddress(ModuleHandle, reinterpret_cast<LPCSTR>(2577)));
+    if (!ProcAddress)
+    {
+        return FALSE;
+    }
+
+    return ProcAddress();
 }